+++ extras_old/helper.go
@@ -1,4 +1,4 @@
-package extras
+package extras_old
 
 import "github.com/yuin/goldmark/util"
