+++ extras_old/insert_test.go
@@ -1,4 +1,4 @@
-package extras
+package extras_old
 
 import (
 	"bytes"
